Device visualizer - Added LedClicked event
Device properties - Added the ability to select LEDs by clicking on them in the preview

+++ src/Artemis.Core/Extensions/RgbDeviceExtensions.cs
@@ -25,10 +25,10 @@ namespace Artemis.Core
         public static SKRect ToSKRect(this Rectangle rectangle)
         {
             return SKRect.Create(
-                (float) rectangle.Location.X,
-                (float) rectangle.Location.Y,
-                (float) rectangle.Size.Width,
-                (float) rectangle.Size.Height
+                rectangle.Location.X,
+                rectangle.Location.Y,
+                rectangle.Size.Width,
+                rectangle.Size.Height
             );
         }
     }
